Implemented RevealActionCard
everything should be implemented for the reveal action card.
Added the class int he TileO model
also added it in the state machine, and in the controllers.
also takes into account if someone tries to reveal more than 1 tile.
Did regeneration through umple for the new class etc.
created a new Pop out window for the new actioncard.
also updated the design interface to add a new type of action card.

the only thing that might be left to do is regenerate the ample code
and make sure everything actually works well and all the code
corresponds to the correct code by umple because i might have changed
some things “illegally” (aka directly in the java code and not through
umple)